+++ src/nodejs/lambda-handlers/notification-consumer/templates.js
@@ -1,3 +1,5 @@
+const db = require('database-util');
+
 const templates = {
   direct_message: (e) => ({
     user_id: e.user_id,
@@ -35,13 +37,48 @@ const templates = {
   })
 };
 
-const getTemplate = (message) => {
+const getTemplate = async (message) => {
   if (message.event_type && templates[message.event_type]) {
     const template = templates[message.event_type](message);
     template.conversation_id = template.conversation_id || message.conversation_id;
+    if (message.step_name && message.event_type !== 'request_initialized') {
+      const stepMessage = await db.submission.getStepMessage({ step_name: message.step_name });
+      template.text = `${template.text}\n${stepMessage.notification}`;
+    }
     return template;
   }
   return false;
 };
 
+const getEmailTemplate = async (eventMessage, message) => {
+  let emailPayload = {};
+  if (eventMessage.event_type !== 'direct_message') {
+    const workflowName = db.workflow.getLongName({ id: eventMessage.workflow_id });
+    const formData = await db.submission.getFormData({ id: eventMessage.submission_id });
+
+    emailPayload = {
+      name: 'EDPUB User',
+      submission_id: eventMessage.submission_id,
+      workflow_name: (await workflowName).long_name,
+      conversation_last_message: message.text
+    };
+
+    if (formData.data_product_name_value) {
+      emailPayload.submission_name = formData.data_product_name_value;
+    } else { (emailPayload.submission_name = `Request Initialized by ${(await db.submission.getCreatorName({ id: eventMessage.submission_id })).name}`); }
+  }
+
+  if (eventMessage.event_type !== 'form_submitted'
+    && eventMessage.event_type !== 'review_approved') {
+    emailPayload.step_name = eventMessage.step_name;
+  } else {
+    emailPayload.step_name = (await db.submission.getStepName({
+      id: eventMessage.submission_id
+    })).step_name;
+  }
+
+  return emailPayload;
+};
+
 module.exports.getTemplate = getTemplate;
+module.exports.getEmailTemplate = getEmailTemplate;