+++ src/nodejs/lambda-layers/database-util/src/query/parsers.js
@@ -70,3 +70,7 @@ module.exports.getConversationId = one;
 module.exports.getFormData = one;
 module.exports.getUsers = many;
 module.exports.setDetail = one;
+module.exports.getStepMessage = one;
+module.exports.getLongName = one;
+module.exports.getCreatorName = one;
+module.exports.getStepName = one;