+++ terraform/aws_secrets/outputs.tf
@@ -1,0 +1,3 @@
+output "ses_access_creds_arn" {
+  value = aws_secretsmanager_secret.ses_access_creds.arn
+}